refactor: LotteryConverter 제거 후, 해당 역할 Main으로 위임
