+++ src/main/java/lotterymachine/utils/LotteryCalculator.java
@@ -1,6 +1,7 @@
 package lotterymachine.utils;
 
-import lotterymachine.Money;
+import lotterymachine.domain.LotteryTicket;
+import lotterymachine.domain.Money;
 
 
 import static java.lang.Math.floor;
@@ -9,7 +10,7 @@ public class LotteryCalculator {
     private static final double DECIMAL_PLACE_SAVER = 100.0;
 
     public static int divideByLotteryPrice(int amount) {
-        return amount / Money.PER_LOTTERY_TICKET_PRICE;
+        return amount / LotteryTicket.PER_PRICE;
     }
 
     public static double calculateProfitRate(double winningLotteryAmount, int amount) {
