+++ src/test/java/lotterymachine/domain/CountTest.java
@@ -1,7 +1,5 @@
 package lotterymachine.domain;
 
-import lotterymachine.Money;
-import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
